chore: clear workspace clippy debt + rustfmt under rust 1.92
Fix the pre-existing clippy findings across the workspace (audio, physics, fluid, spatial, noise, bytes, crossdomain, etc.) that were failing `cargo clippy --workspace --all-targets -D warnings`, and apply rustfmt under the new toolchain. Mechanical, no behavior change intended; full test suite stays green (2328 passing). Includes refactoring an over-long arg list into a params struct (audio ImageSourceContext), removing a dead recursion param (spatial kdtree), and rewriting an out-of-date bench (unshape-noise) to the current struct API.

+++ crates/unshape-audio/src/midi.rs
@@ -488,7 +488,7 @@ pub fn parse_note_name(name: &str) -> Option<Note> {
     };
 
     let octave: i8 = octave_part.parse().ok()?;
-    let note = ((octave + 1) as i16 * 12 + base_note as i16) as i16;
+    let note = (octave + 1) as i16 * 12 + base_note as i16;
 
     if (0..=127).contains(&note) {
         Some(note as Note)
